Moves sysmon proc node to nodes
