+++ beagle/nodes/__init__.py
@@ -5,7 +5,7 @@ from .domain import URI, Domain
 from .file import File, FileOf
 from .ip_address import IPAddress
 from .node import Node
-from .process import Launched, Process
+from .process import Process, SysMonProc
 from .registry import RegistryKey
 
 
@@ -16,7 +16,7 @@ __all__ = [
     "File",
     "FileOf",
     "IPAddress",
-    "Launched",
+    "SysMonProc",
     "Process",
     "RegistryKey",
     "Alert",
